move computed rezoning props to model and cleanup relics in model
